getting rid of more references
